rearrange some files
